`MonadFix` can only be lifted through `MonadTransControlIdentity`.
The type checker also accepted the old version, but unfortunately,
whenever there was monadic state involved, the program would hang.

I am not 100% sure, whether this really is impossible to just handle it
with laziness annotations.

But then again for `ExceptT`, there would have to be an additional case
handling `Left`.

But I can safely say, the previous implementation here hung for
`StateT` and `ExceptT`, when the implementations from "containers" still
worked.