Add `Control.Monad.Trans.Compose.Stack` and `Control.Monad.Trans.Compose.Transparent`.
`Control.Monad.Trans.Compose.Stack` implements a simple interface, which
can be used define and run monad transformer stacks.

The main motivation here, is that the very first transformer in a
transformer stack built with `ComposeT` doesn't add the expected
instances to the stack.

The solution is, to use an additional `TransparentT` with every
transformer stack.

This interface makes the unexpected behaviour impossible.