Update mtl to 2.3.
Implement `Elevator` instances for `MonadAccum` and `MonadSelect`.
Implement recursive `ComposeT` instances for `MonadAccum` and `MonadSelect`.

The base-case `ComposeT` instances are a bit weird.
`AccumT` and `SelectT` are expected to be at the bottom of a monad
transformer stack.

The only idea I had until now, was abusing a
`MonadBaseControlIdentity Identity` constraint.
The problem with this approach is, that this doesn't cover `((->) r)` as
a base monad.
Maybe this isn't too bad, because in `mtl` this isn't supported anyways
at the moment.